refactor: moved claims to base client response with errors
