+++ client/src/main/java/io/jans/as/client/UserInfoResponse.java
@@ -9,18 +9,12 @@ package io.jans.as.client;
 import io.jans.as.model.userinfo.UserInfoErrorResponseType;
 import org.jboss.resteasy.client.ClientResponse;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 /**
  * Represents an user info response received from the authorization server.
  *
  * @author Javier Rojas Blum Date: 11.30.2011
  */
 public class UserInfoResponse extends BaseResponseWithErrors<UserInfoErrorResponseType> {
-
-    private Map<String, List<String>> claims;
 
     /**
      * Constructs a User Info response.
@@ -29,28 +23,11 @@ public class UserInfoResponse extends BaseResponseWithErrors<UserInfoErrorRespon
      */
     public UserInfoResponse(ClientResponse<String> clientResponse) {
         super(clientResponse);
-        claims = new HashMap<>();
     }
 
     @Override
     public UserInfoErrorResponseType fromString(String str) {
         return UserInfoErrorResponseType.fromString(str);
-    }
-
-    public Map<String, List<String>> getClaims() {
-        return claims;
-    }
-
-    public void setClaims(Map<String, List<String>> claims) {
-        this.claims = claims;
-    }
-
-    public List<String> getClaim(String claimName) {
-        if (claims.containsKey(claimName)) {
-            return claims.get(claimName);
-        }
-
-        return null;
     }
 
     @Override
@@ -59,7 +36,6 @@ public class UserInfoResponse extends BaseResponseWithErrors<UserInfoErrorRespon
                 "status=" + status +
                 "entity=" + entity +
                 "headers=" + headers +
-                "claims=" + claims +
                 ", super=" + super.toString() +
                 '}';
     }
